requirements.txt
requests>=2.25.0  # For label downloading
prometheus-client>=0.11.0  # For metrics
itsdangerous>=2.1.0
fastapi
pillow
numpy
tensorflow
uvicorn
python-multipart
